requirements.txt
asgiref==3.7.2
asttokens==2.2.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
backcall==0.2.0
certifi==2023.11.17
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
colorama==0.4.6
constantly==23.10.4
crispy-tailwind==1.0.3
cryptography==43.0.0
daphne==4.1.2
decorator==5.1.1
Django==4.2.7
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-extensions==3.2.3
et-xmlfile==1.1.0
executing==1.2.0
hyperlink==21.0.0
gunicorn
idna==3.6
incremental==22.10.0
ipython==8.14.0
jedi==0.18.2
line_notify==0.1.4
matplotlib-inline==0.1.6
mysqlclient==2.2.1
openpyxl==3.1.2
parso==0.8.3
pickleshare==0.7.5
Pillow==10.1.0
prompt-toolkit==3.0.38
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
Pygments==2.15.1
pyOpenSSL==24.2.1
psycopg2-binary
requests==2.31.0
service-identity==24.1.0
setuptools==69.5.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.9.0
Twisted==24.3.0
twisted-iocpsupport==1.0.4; platform_system == "Windows"
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.6
wheel==0.43.0
zope.interface==6.4.post2
django-environ
psycopg2-binary
